Criação da separação das rotas da API e disponibilização do .env pra conectar ao banco de dados
