+++ api/.env
@@ -1,0 +1,8 @@
+# Configurações do Servidor
+PORT=3000
+
+# Configurações do Banco de Dados
+DB_HOST=localhost
+DB_USER=root
+DB_PASSWORD=
+DB_NAME=minha_api_db
