+++ api/.gitignore
@@ -72,12 +72,14 @@ web_modules/
 # Yarn Integrity file
 .yarn-integrity
 
+#deixei os .env comentado para nao dar problema com o arquivo com dados pra conectar no banco local
+
 # dotenv environment variable files
-.env
-.env.development.local
-.env.test.local
-.env.production.local
-.env.local
+#.env
+#.env.development.local
+#.env.test.local
+#.env.production.local
+#.env.local
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
